Refactor relevance check (#112)
Another refactor to make answer_relevant an actual property. When called, it will trigger a validation automatically.
This is just better UX. The user needs to otherwise be aware of the fact that buster.postprocess_completion needs to be called. This PR will now:

* explicitly runs all postprocessing computations at the end of a generation

* renames variables

* accepts answer_text directly as input for a Completion

In short, after the text is generated in answer_generator, a postprocess function is automatically executed, computing if necessary answer relevance as well as re-ranking documents